added login with email
