+++ src/main/java/org/launchcode/professionalprocrastinators/models/dto/LoginFormDTO.java
@@ -15,14 +15,6 @@ public class LoginFormDTO {
     @Size(min = 3, max = 20, message = "Invalid username. Must be between 3 and 20 characters.")
     private String username;
 
-    @NotEmpty(message = "Can not be empty")
-    @NotBlank(message = "Name is required")
-    private String name;
-
-    @NotEmpty(message = "Can not be empty")
-    @NotBlank(message = "Username is required")
-    @Email(message = "Must be a valid email")
-    private String email;
 
     @NotNull
     @NotBlank
@@ -35,22 +27,6 @@ public class LoginFormDTO {
 
     public void setUsername(String username) {
         this.username = username;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
     }
 
     public String getPassword() {
