+++ src/main/java/org/launchcode/professionalprocrastinators/models/dto/RegisterFormDTO.java
@@ -1,6 +1,8 @@
 package org.launchcode.professionalprocrastinators.models.dto;
 
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Size;
 import org.jetbrains.annotations.NotNull;
 
@@ -10,6 +12,32 @@ public class RegisterFormDTO extends LoginFormDTO {
     @NotBlank
     @Size(min = 5, max = 30, message = "Invalid password. Must be between 5 and 30 characters.")
     private String verifyPassword;
+
+    @NotEmpty(message = "Can not be empty")
+    @NotBlank(message = "Name is required")
+    private String name;
+
+    @NotBlank(message = "Username is required")
+    @Email(message = "Must be a valid email")
+    private String email;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
 
     public String getVerifyPassword() {
         return verifyPassword;
