+++ src/main/resources/templates/login.html
@@ -10,7 +10,7 @@
 
 <form method="post">
     <div class="form-group">
-        <label th:for="username">Username
+        <label th:for="username">Username or Email
             <input class="form-control" th:field="${loginFormDTO.username}">
         </label>
         <p class="error" th:errors="${loginFormDTO.username}"></p>
